requirements.txt
# Core scientific computing
numpy>=1.26.0
scipy>=1.14.0
pandas>=2.2.0
xarray>=2024.3.0
matplotlib>=3.9.0

# Utilities
requests>=2.31.0  # API calls for ingestion
pyyaml>=6.0.2     # Config loading
click>=8.1.0      # CLI
tenacity>=8.2.0   # Retries
pyproj>=3.6.0     # Coordinate transformations

# Parallel and performance
dask[complete]>=2024.8.0
joblib>=1.4.0

# Database and storage
sqlalchemy>=2.0.0
h5py>=3.11.0

# Geospatial (core, GDAL system dep recommended via conda)
geopandas>=1.0.0
rasterio>=1.3.10

# Optional: Geophysics (pip install .[geophysics])
# simpeg>=0.21.0
# obspy>=1.4.1
# mintpy>=1.6.0
# pygimli>=1.5.0

# Optional: Visualization (pip install .[visualization])
# pygmt>=0.12.0
# pyvista>=0.44.0
# folium>=0.17.0
# vtk>=9.3.0

# Note: GDAL/GEOS via system or conda for geospatial; not pinned in pip due to binaries

# Development/Testing (pip install -e .[dev])
# pytest>=8.3.0
# black>=24.4.2
# flake8>=7.0.0
# coverage>=7.6.0
# sphinx>=7.4.0
# psutil>=6.0.0